pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "protllm"
version = "1.0.0"
description = "An Interleaved Protein-Language LLM with Protein-as-Word Pre-Training"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch==2.1.1", "transformers==4.36.2", "sentencepiece==0.1.99", "tokenizers==0.15.0",
    "accelerate==0.25.0", "peft==0.7.0", "numpy", "scikit-learn==1.3.2", "wandb",
    "fair-esm"
]

[project.optional-dependencies]
build = ["build", "twine"]

[project.urls]
"Homepage" = "https://protllm.github.io/project"
"Bug Tracker" = "https://github.com/ProtLLM/ProtLLM/issues"

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]